Move provider's log task handlers to the provider package (#9604)
