+++ tests/providers/elasticsearch/log/elasticmock/__init__.py
