+++ tests/providers/elasticsearch/log/elasticmock/fake_elasticsearch.py
