+++ tests/providers/elasticsearch/log/elasticmock/utilities/__init__.py
