+++ tests/providers/elasticsearch/log/test_es_task_handler.py
@@ -30,8 +30,8 @@ from parameterized import parameterized
 from airflow.configuration import conf
 from airflow.models import DAG, TaskInstance
 from airflow.operators.dummy_operator import DummyOperator
+from airflow.providers.elasticsearch.log.es_task_handler import ElasticsearchTaskHandler
 from airflow.utils import timezone
-from airflow.utils.log.es_task_handler import ElasticsearchTaskHandler
 from airflow.utils.state import State
 from airflow.utils.timezone import datetime
